compute-api/requirements.txt
Flask
psutil
pyjwt<2.0
cryptography
gunicorn
flask_sqlalchemy
psycopg2-binary>=2.8
python-dotenv
flask-cors
rq
redis